requirements.txt
Flask>=2.0
requests>=2.25
openai==1.82.0
anthropic==0.49.0
google-generativeai==0.8.5 